Restructure project for PyPI publishing with src/ layout
MAJOR RESTRUCTURING:

📦 Package Structure:
- Moved agentgrunt/ → src/agor/ (proper src layout)
- Renamed gpt_tools/ → tools/ (cleaner naming)
- Added proper __init__.py with package metadata

🔧 Build System:
- Replaced poetry with setuptools build system
- Added setup.cfg with complete package configuration
- Updated pyproject.toml to use setuptools backend
- Added MANIFEST.in for including non-Python files
- Removed setup.py (using setup.cfg instead)

🚀 PyPI Publishing:
- Added GitHub workflow for automated PyPI publishing
- Configured for both manual dispatch and release triggers
- Uses trusted publishing with OIDC tokens
- Environment protection for release security

📝 Documentation Updates:
- Updated all file paths in documentation
- Fixed import statements in tests
- Updated cache directory from 'agentgrunt' to 'agor'
- Corrected all tool references to new structure

🎯 Entry Points:
- Maintained both 'agor' and 'agentgrunt' commands
- Updated console scripts in setup.cfg
- Preserved backward compatibility

BENEFITS:
- Standard Python src/ layout for better packaging
- Automated PyPI publishing workflow
- Cleaner project structure
- Proper package metadata and dependencies
- Ready for pip install from PyPI
