+++ setup.cfg
@@ -1,0 +1,48 @@
+[metadata]
+name = agor
+version = 1.0.0
+author = Jeremiah K.
+author_email = jeremiahk@gmx.com
+description = AgentOrchestrator - Multi-Agent Development Coordination Platform
+long_description = file: AGOR_README.md
+long_description_content_type = text/markdown
+url = https://github.com/jeremiah-k/agor
+project_urls =
+    Bug Tracker = https://github.com/jeremiah-k/agor/issues
+    Repository = https://github.com/jeremiah-k/agor
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+    Development Status :: 4 - Beta
+    Topic :: Software Development :: Libraries :: Python Modules
+    Topic :: Software Development :: Quality Assurance
+    Intended Audience :: Developers
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.11
+install_requires =
+    fastapi
+    httpx
+    plumbum
+    pydantic
+    pyperclip
+    pytest
+    tqdm
+    typer
+    platformdirs>=4.3.7,<5.0.0
+include_package_data = True
+
+[options.packages.find]
+where = src
+
+[options.package_data]
+agor.tools = *.md, *.py
+
+[options.entry_points]
+console_scripts =
+    agor = agor.main:app
+    agentgrunt = agor.main:app
