+++ src/agor/__init__.py
@@ -1,0 +1,11 @@
+"""
+AgentOrchestrator (AGOR) - Multi-Agent Development Coordination Platform
+
+A comprehensive project planning and multi-agent coordination platform.
+Plan complex development projects, design agent teams, and generate 
+specialized prompts for coordinated AI development workflows.
+"""
+
+__version__ = "1.0.0"
+__author__ = "Jeremiah K."
+__email__ = "jeremiahk@gmx.com"
