+++ src/agor/tools/README_ai.md
